+++ src/pages/StudentDetails.js
@@ -11,7 +11,10 @@ export const StudentDetails = () => {
     }, [])
 
     const deleteAccount = () => {
-        axios.delete("http://localhost:8080/delete-student-account/" + student._id).then(() => navigate("/"))
+        const isConfirm = window.confirm("Are you sure?")
+        if (isConfirm) {
+            axios.delete("http://localhost:8080/delete-student-account/" + student._id).then(() => navigate("/"))
+        }
     }
     return (
         <div className='flex flex-col items-center mt-2 text-[14px] lg:text-[16px]'>
